pyproject.toml
[project]
name = "lubabot"
version = "0.1.0"
description = "Telegram booking bot with Google Sheets mirror"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.4",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.19",
    "gspread>=6.0",
    "google-auth>=2.28",
    "apscheduler>=3.10",
    "tenacity>=8.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3",
    "mypy>=1.8",
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "ASYNC"]

[tool.mypy]
python_version = "3.11"
strict = true
plugins = ["pydantic.mypy"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
